+++ src/main/java/com/github/junrar/impl/FileVolumeManager.java
@@ -13,21 +13,14 @@ import java.io.File;
  */
 public class FileVolumeManager implements VolumeManager {
     private final File firstVolume;
-    private final String password;
-
-    public FileVolumeManager(final File firstVolume, final String password) {
-        this.firstVolume = firstVolume;
-        this.password = password;
-    }
 
     public FileVolumeManager(final File firstVolume) {
         this.firstVolume = firstVolume;
-        this.password = null;
     }
 
     @Override
     public Volume nextArchive(final Archive archive, final Volume last) {
-        if (last == null) return new FileVolume(archive, this.firstVolume, password);
+        if (last == null) return new FileVolume(archive, this.firstVolume);
 
         final FileVolume lastFileVolume = (FileVolume) last;
         final boolean oldNumbering = !archive.getMainHeader().isNewNumbering()
@@ -35,6 +28,6 @@ public class FileVolumeManager implements VolumeManager {
         final String nextName = VolumeHelper.nextVolumeName(lastFileVolume.getFile().getAbsolutePath(), oldNumbering);
         final File nextVolume = new File(nextName);
 
-        return new FileVolume(archive, nextVolume, password);
+        return new FileVolume(archive, nextVolume);
     }
 }
